+++ src/Pim/Bundle/EnrichBundle/Form/Type/AttributeOptionType.php
@@ -22,15 +22,9 @@ class AttributeOptionType extends AbstractType
     {
         $this->addFieldId($builder);
 
-        $this->addFieldSortOrder($builder);
-
-        $this->addFieldTranslatable($builder);
-
         $this->addFieldOptionValues($builder);
 
         $this->addFieldCode($builder);
-
-        $this->addFieldIsDefault($builder);
     }
 
     /**
@@ -40,24 +34,6 @@ class AttributeOptionType extends AbstractType
     protected function addFieldId(FormBuilderInterface $builder)
     {
         $builder->add('id', 'hidden');
-    }
-
-    /**
-     * Add field sort_order to form builder
-     * @param FormBuilderInterface $builder
-     */
-    protected function addFieldSortOrder(FormBuilderInterface $builder)
-    {
-        $builder->add('sort_order', 'integer', array('required' => false));
-    }
-
-    /**
-     * Add field translatable to form builder
-     * @param FormBuilderInterface $builder
-     */
-    protected function addFieldTranslatable(FormBuilderInterface $builder)
-    {
-        $builder->add('translatable', null, array('required' => false));
     }
 
     /**
@@ -85,15 +61,6 @@ class AttributeOptionType extends AbstractType
                 'by_reference' => false,
             )
         );
-    }
-
-    /**
-     * Add isDefault field to form builder
-     * @param FormBuilderInterface $builder
-     */
-    protected function addFieldIsDefault(FormBuilderInterface $builder)
-    {
-        $builder->add('default', 'hidden');
     }
 
     /**