+++ src/Pim/Bundle/UIBundle/Resources/public/js/pim-formupdatelistener.js
@@ -14,7 +14,6 @@ define(
                 self  = this,
                 formUpdated = function (e) {
                     var $target = $(e.target);
-
                     if ($target.parents('div.filter-box').length || $target.parents('ul.icons-holder').length || $target.hasClass('exclude')) {
                         return;
                     }